+++ srv/option/sort.go
@@ -18,7 +18,7 @@ type SortQueryParser interface {
 // ResolvesSortQuery parses the sort parameters from the URL query and
 // stores them within the JSON:API context.
 func ResolvesSortQuery(parser SortQueryParser) srv.Options {
-	return srv.UseMiddleware(func(next http.Handler) http.Handler {
+	return srv.WithMiddleware(func(next http.Handler) http.Handler {
 		return sortParamResolver{next, parser}
 	})
 }
